moved styled components to shared folder
