+++ src/components/Layout/index.jsx
@@ -1,6 +1,6 @@
 import { NavBar } from "../NavBar/index.jsx";
 import { Footer } from "../Footer/index.jsx";
-import { VerticalContainer } from "../styles/verticalContainer.styled.js";
+import { VerticalContainer } from "../../shared/styledComponents/verticalContainer.styled.js";
 
 export function Layout({ children }) {
   return (
