+++ src/pages/Results/index.jsx
@@ -1,6 +1,6 @@
 import { Banner } from "../../components/Banner/index.jsx";
 import { ResultsBody } from "../../components/ResultsBody/index.jsx";
-import { PageWidthContainer } from "../../components/styles/pageWidthContainer.js";
+import { PageWidthContainer } from "../../shared/styledComponents/pageWidthContainer.js";
 
 export function Results() {
   return (
