+++ src/shared/styledComponents/gridElement.styled.js
