+++ src/shared/styledComponents/rectangularFrame.styled.js
