+++ src/shared/styledComponents/verticalContainer.styled.js
